Clousures are working correctly now!
The solution was to save a `Rc<RefCell<Env>>` at the time of a
lambda's creation referencing the current environment. Then, when
we try to apply this lambda to a value, we create a local environment
for its parameter and local definitions, and then link the closure to
this environment so that if a name lookup fails in the local
environment, we check the next (remembered) one.

Moreover, the `Env` struc was split into an `Env` and a `Cont` since
the `loops` variable is a seperate concern. But also because there
were issues with dynamic borrowing because of it.

Thanks to `Rc<RefCell<T>>`, it's possible for multiple closures to
keep an (eventually mutable) reference to one Env, so they can all
access it during the host program's execution dynamically. In practice,
only one closure will need a .borrow_mut() to the `Env` at a given time,
as we have no parallelism features yet!