Nearly 2x faster: compiling with closures.
The first problem with the old interpreter was that there was no
notion of a `Value` seperate from the AST's `Expr`. This made the
logic needlessly complicated and tangled together.

Secondly, all execution was tied to the ast with no IR of any kind
to provide a way of optimising the hot parts. This is solved with
dynamic dispatch of arbitrary closures that represent the code of
an `Expr` or an `Instr`; this was, the interaction with the AST is
only done once to produce a closure capable of returning a `Value`.
In the case of evaluted functions, their Instr body contains "compiled"
versions of the original instructions.

fnv is noticeably faster than rustc_hash for our purposes, so
that's what we use now.