[MLIR] safe API for attributes, blocks, regions, contexts and values!
More of the unsafe C API is wrapped for convenient use, while the
bindings are constrained to the chimera::mlir::raw module.

The mod structure follows the core IR components i.e: attributes,
blocks, operations, regions, types and values, making each of them
into a separate module.

The methods .as_raw() and from_raw() are used to convert between
the C API structs and the higher-level Rust API structs. This may
deserve a AsRaw and FromRaw traits in the future.