[MLIR] safe API for types, passes and locations!
MLIR types now have their own module, they're implement as Copy types
since the underlying memory is handled by the context anyway.

There is some doubt about the lifetimes of Type, it may possible to drop
the Context while still holding a type, this could be solved by
requiring all use of the type to have a reference to the context.

The PassManager is wrapped in a builder fashion with $dialect_to_$dialect
methods that add passes to it.