[MLIR] fix lifetime issues with non-owning pointers.
Up until this point, values such as Type, Attribute and Location
did not own the data they point to, the context (or similar) they
were created in was the owner, this allowed us to "leak" data from
Context or Operation and essential pass dangling pointers around.

Enter PhantomData, with this we can force these tricky types to own
references to their owning objects, so that the compiler's lifetime
analysis can keep track of their validity and complain when necessary.

The from_raw function is not unsafe since it can allow for building
values with inconsistent lifetimes, i.e not associated with its owner.
This information is only available as an object model specified by the
C API and isn't something the compiler can or should be aware of.