[CST] add new intermediate representation produced by the typechecker.
The CST (Concrete Syntax Tree) is a new version of the AST with type
annotations for all expressions. Moreover, we have sufficient
information at this stage to know if lambda applications are staturated,
so we can peform fast C-like calls.

The typechecker now transforms each AST expression into a CST typed
expression and drops location information in the process. This part
might change in the future.