[AST] reduce the AST to the typed lambda calculus for prototyping.
We also added location information for all expressions and items
using a generic Node struct.

Moreover, the AST no longer represents lambdas and applications as nested
recursive data structures, i.e an expression of the form `((f 1) 2) 3`
is represented as `(f, [1, 2, 3])`. This makes the grammar simpler and
permits easier handling of the data structure later on.

The lexer no longer depends on phf, we instead use match expressions
to provide the same functionality. This should provide speedups in
compilation and eliminate the bothersome macro expansion errors.